chore: renaming to ServiceFactory
