+++ packages/backend-app-api/src/services/implementations/cache/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { cacheFactory } from './cacheFactory';
+export { cacheServiceFactory } from './cacheServiceFactory';
