+++ packages/backend-app-api/src/services/implementations/config/index.ts
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 
-export { configFactory } from './configFactory';
-export type { ConfigFactoryOptions } from './configFactory';
+export { configServiceFactory } from './configServiceFactory';
+export type { ConfigFactoryOptions } from './configServiceFactory';
