+++ packages/backend-app-api/src/services/implementations/logger/loggerServiceFactory.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,4 +14,19 @@
  * limitations under the License.
  */
 
-export { rootLoggerServiceFactory } from './rootLoggerServiceFactory';
+import {
+  createServiceFactory,
+  coreServices,
+} from '@backstage/backend-plugin-api';
+
+/** @public */
+export const loggerServiceFactory = createServiceFactory({
+  service: coreServices.logger,
+  deps: {
+    rootLogger: coreServices.rootLogger,
+    plugin: coreServices.pluginMetadata,
+  },
+  factory({ rootLogger, plugin }) {
+    return rootLogger.child({ plugin: plugin.getId() });
+  },
+});
