+++ packages/backend-app-api/src/services/implementations/rootHttpRouter/index.ts
@@ -15,10 +15,10 @@
  */
 
 export {
-  rootHttpRouterFactory,
+  rootHttpRouterServiceFactory,
   type RootHttpRouterFactoryOptions,
   type RootHttpRouterConfigureOptions,
-} from './rootHttpRouterFactory';
+} from './rootHttpRouterServiceFactory';
 export {
   DefaultRootHttpRouter,
   type DefaultRootHttpRouterOptions,
