+++ packages/backend-app-api/src/services/implementations/tokenManager/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { tokenManagerFactory } from './tokenManagerFactory';
+export { tokenManagerServiceFactory } from './tokenManagerServiceFactory';
