+++ packages/backend-app-api/src/services/implementations/urlReader/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { urlReaderFactory } from './urlReaderFactory';
+export { urlReaderServiceFactory } from './urlReaderServiceFactory';
